+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.13.0
 * Added the option to download the dataset file.
 * Added create folder, delete and collapse icons in tree explorer
+* Added the option to submit the JCL
 
 ## 0.12.0
 * Added the option to open a file in new browser tab